Refactorment for JobSearch components. Moved them to a seperate folder
