+++ src/App.vue
@@ -5,7 +5,7 @@
 
 <script>
 import MainNav from "@/components/Navigation/MainNav.vue";
-import MainHero from "@/components/MainHero.vue";
+import MainHero from "@/components/JobSearch/MainHero.vue";
 
 export default {
   name: "App",
