+++ tests/unit/components/JobSearch/TheHeadline.test.js
@@ -1,7 +1,6 @@
 import { nextTick } from "vue";
 import { render, screen } from "@testing-library/vue";
-import TheHeadline from "@/components/TheHeadline.vue";
-import { vi } from "vitest";
+import TheHeadline from "@/components/JobSearch/TheHeadline.vue";
 
 describe("TheHeadline", () => {
   beforeEach(() => {
